@@ -2,22 +2,20 @@
   "name": "varside",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "element-ui": "^2.15.9",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.1"
+    "vue": "^3.2.37",
+    "vue-router": "^4.1.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "sass": "^1.32.7",
-    "sass-loader": "^12.0.0",
-    "vue-template-compiler": "^2.6.14"
+    "@vitejs/plugin-vue": "^3.0.0",
+    "sass": "^1.54.3",
+    "sass-loader": "^13.0.2",
+    "vite": "^3.0.0"
   }
 }
