@@ -1,15 +1,5 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
+import './style.css'
 import App from './App.vue'
-import router from './router'
 
-Vue.config.productionTip = false
-
-import ElementUI from 'element-ui'
-import 'element-ui/lib/theme-chalk/index.css'
-
-Vue.use(ElementUI)
-
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).mount('#app')
